Source Term ID Bug Fixes
Fixes bugs pertaining to adding a Source Term ID to the tagged terms. It does this by implementing an extra data member and functions to the TaggedTerms object. There is also a warning if source_terms_ids is not empty but is not used due to unequal length.
Also exposes the TaggedTerm object on the top layer of t2t.