Fixes to Clarity and Terminology
Adds comments to the outputted csv files for text2term when the save_mappings options in enabled. These comments include information on the mapper and settings used, as well as the time and version that produced the results.
Also changed all references from 'blacklist' to 'blocklist' to use modern terminology in the preprocessing. Still allows for backwards compatibility.